moved test to many-to-many.test.ts
